Make several updates
New argument 'bm' for a benchmark series:
if provided, the ratio of series is computed
(see https://ssrn.com/abstract=3374195). Also,
annualised returns are the differences between
the annualised returns of the orignal series
and the benchmark's annualised return.

New argument 'currency', with default "USD". Only
used for the labels.

New argument 'arrow'. Only used for labels.

Several fixes to the axes labels.